corretto errore (ho erroneamente considerato un componente come funzione)
